* Fixed rusty remover not being removed properly * Changed getLevel/magicLevel functions to unsigned integers * Fixed spells not being loaded without "maglv" set * Configmanager doesn't need to load the lua libraries

+++ otserv/trunk/luascript.cpp
@@ -2455,7 +2455,7 @@ int LuaScriptInterface::luaDoPlayerRemoveItem(lua_State *L)
 int LuaScriptInterface::luaDoFeedPlayer(lua_State *L)
 {
 	//doFeedPlayer(uid, food)
-	int32_t food = (int32_t)popNumber(L);
+	uint32_t food = popNumber(L);
 	uint32_t cid = popNumber(L);
 
 	ScriptEnviroment* env = getScriptEnv();
